+++ src/camera.h
@@ -12,23 +12,23 @@ namespace Chotra_RT {
     class Camera {
     public:
         Camera();
-        float GetViewportWidth() const;
-        float GetViewportHeight() const;
-        float GetFocalLength() const;
+        double GetViewportWidth() const;
+        double GetViewportHeight() const;
+        double GetFocalLength() const;
 
-        glm::vec3 GetCameraCenter() const;
+        glm::dvec3 GetCameraCenter() const;
 
     private:
 
-        float focal_length_ = 1.0f;
-        float viewport_height_ = 2.0f;
-        float viewport_width_ = 2.0f;
+        double focal_length_ = 1.0f;
+        double viewport_height_ = 2.0f;
+        double viewport_width_ = 2.0f;
         
-        glm::vec3 camera_center_ = glm::vec3(0.0f, 0.0f, 0.0f);
+        glm::dvec3 camera_center_ = glm::dvec3(0.0f, 0.0f, 0.0f);
 
-        glm::vec3 front_ = glm::vec3(0.0f, 0.0f, -1.0f);
-        glm::vec3 up_ = glm::vec3(0.0f, 1.0f, 0.0f);
-        glm::vec3 right_ = glm::vec3(1.0f, 0.0f, 0.0f);
+        glm::dvec3 front_ = glm::dvec3(0.0f, 0.0f, -1.0f);
+        glm::dvec3 up_ = glm::dvec3(0.0f, 1.0f, 0.0f);
+        glm::dvec3 right_ = glm::dvec3(1.0f, 0.0f, 0.0f);
 
     };
