+++ src/ray.cpp
@@ -2,20 +2,20 @@
 
 namespace Chotra_RT {
 
-    Ray::Ray(const glm::vec3 origin, const glm::vec3 direction) 
+    Ray::Ray(const glm::dvec3 origin, const glm::dvec3 direction) 
             : origin_(origin), direction_(direction) {
 
     }
 
-    glm::vec3 Ray::GetOrigin() const {
+    glm::dvec3 Ray::GetOrigin() const {
         return origin_;
     }
 
-    glm::vec3 Ray::GetDirection() const {
+    glm::dvec3 Ray::GetDirection() const {
         return direction_;
     }
 
-    glm::vec3 Ray::Func(float t) const {
+    glm::dvec3 Ray::Func(double t) const {
         return origin_ + t * direction_;
     }
